requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
pillow==11.2.1
chromadb==0.6.3
sentence-transformers==4.1.0
google-genai==1.70.0
